requirements.txt
coverage==4.3.4
Django==1.10.5
djangorestframework==3.5.3
psycopg2==2.6.2
